py/requirements.txt
  
google
requests
-e git+https://github.com/jstasiak/python-zeroconf.git#egg=python-zeroconf
pycryptodome
diffiehellman
pyaes
protobuf
